requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via starlette
async-timeout==4.0.3
    # via redis
billiard==4.2.1
    # via celery
celery==5.4.0
    # via
    #   -r requirements.in
    #   flower
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.0
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
humanize==4.11.0
    # via flower
idna==3.10
    # via anyio
kombu==5.4.2
    # via celery
loguru==0.7.2
    # via -r requirements.in
prometheus-client==0.21.0
    # via flower
prompt-toolkit==3.0.48
    # via click-repl
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.12
    # via -r requirements.in
pytz==2024.2
    # via flower
redis==5.1.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.38.6
    # via fastapi
tornado==6.4.1
    # via flower
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
tzdata==2024.2
    # via
    #   celery
    #   kombu
uvicorn==0.31.1
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
